requirements.txt
# Organ Aging Analysis - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Gradient Boosting (choose one or both)
lightgbm>=3.3.0        # Fast, recommended
xgboost>=1.5.0         # Alternative

# Data Processing
pyarrow>=6.0.0         # For Parquet file support
pyyaml>=6.0            # For configuration files

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model Explainability
shap>=0.41.0           # SHAP values for interpretability

# Dimensionality Reduction & Clustering
umap-learn>=0.5.0      # UMAP for non-linear dimensionality reduction
hdbscan>=0.8.0         # Density-based clustering (optional)

# Jupyter & Notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Utilities
tqdm>=4.62.0           # Progress bars
joblib>=1.1.0          # Model serialization

# Testing (TDD)
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional: Interactive Plotting
plotly>=5.3.0          # Interactive visualizations (optional)

# Optional: Statistical Analysis
statsmodels>=0.13.0    # Advanced statistical models (optional)

# Note: Some packages may have additional system dependencies
# - SHAP may require C++ compiler for optimal performance
# - UMAP benefits from numba (installed automatically)
# - For GPU acceleration: cuml (RAPIDS) can replace UMAP/clustering
